+++ CHANGELOG.md
@@ -1,12 +1,24 @@
 # Changelog
 
+## [19.0.0](https://github.com/luciqai/luciq-reactnative-sdk/compare/v19.0.0...dev)
+
+### Added
+
+- Add support for Screen Rendering ([#4](https://github.com/luciqai/luciq-reactnative-sdk/pull/4)).
+
+### Changed
+
+- Bump Luciq iOS SDK to v19.2.0 ([#4](https://github.com/luciqai/luciq-reactnative-sdk/pull/4)). [See release notes](https://github.com/luciqai/Luciq-iOS-sdk/releases/tag/19.2.0).
+
+- Bump Luciq Android SDK to v19.0.0 ([#4](https://github.com/luciqai/luciq-reactnative-sdk/pull/4)). [See release notes](https://github.com/luciqai/Luciq-Android-sdk/releases/tag/v19.0.0).
+
 ## [18.2.0](https://github.com/luciqai/luciq-reactnative-sdk/compare/v18.2.0...18.0.1)
 
 ### Changed
 
-- Bump Instabug iOS SDK to v18.2.0 ([#11](https://github.com/luciqai/luciq-reactnative-sdk/pull/11)). [See release notes](https://github.com/luciqai/Luciq-iOS-sdk/releases/tag/18.2.0).
+- Bump Luciq iOS SDK to v18.2.0 ([#11](https://github.com/luciqai/luciq-reactnative-sdk/pull/11)). [See release notes](https://github.com/luciqai/Luciq-iOS-sdk/releases/tag/18.2.0).
 
-- Bump Instabug Android SDK to v19.1.0 ([#11](https://github.com/luciqai/luciq-reactnative-sdk/pull/11)). [See release notes](https://github.com/luciqai/Luciq-Android-sdk/releases/tag/v19.1.0).
+- Bump Luciq Android SDK to v18.1.0 ([#11](https://github.com/luciqai/luciq-reactnative-sdk/pull/11)). [See release notes](https://github.com/luciqai/Luciq-Android-sdk/releases/tag/v19.1.0).
 
 ### Added
 
@@ -22,9 +34,9 @@
 
 ### Changed
 
-- Bump Instabug iOS SDK to v18.0.1 ([#6](https://github.com/luciqai/luciq-reactnative-sdk/pull/7)). [See release notes](https://github.com/luciqai/Luciq-iOS-sdk/releases/tag/18.0.1).
+- Bump Luciq iOS SDK to v18.0.1 ([#6](https://github.com/luciqai/luciq-reactnative-sdk/pull/7)). [See release notes](https://github.com/luciqai/Luciq-iOS-sdk/releases/tag/18.0.1).
 
-- Bump Instabug Android SDK to v18.0.1 ([#6](https://github.com/luciqai/luciq-reactnative-sdk/pull/7)). [See release notes](https://github.com/luciqai/Luciq-Android-sdk/releases/tag/v18.0.1).
+- Bump Luciq Android SDK to v18.0.1 ([#6](https://github.com/luciqai/luciq-reactnative-sdk/pull/7)). [See release notes](https://github.com/luciqai/Luciq-Android-sdk/releases/tag/v18.0.1).
 
 ### Fixed
 
@@ -32,4 +44,4 @@
 
 ## [18.0.0](https://github.com/luciqai/luciq-reactnative-sdk/compare/v18.0.0...dev) (September 24, 2025)
 
-- SDK rebranded from Instabug to Luciq.
+- SDK rebranded from Luciq to Luciq.